Removed noise_transform, placed the add_noise function into the image_processing file instead. Moved adversarial cats into a media folder.
